// src/app/modules/support/models/survey.model.ts
export interface ISurveyModel {
  createUserId: string;
  typeNameAz: string;
  updateDate: string;
  updateUserId: string;
  statusNameAz: string;
  typeName: string;
  active: string;
  createUserPhotoFileId: string;
  content: string;
  typeNameEn: string;
  typeNameRu: string;
  typeShortName: string;
  typeColor: string;
  statusColor: string;
  statusId: string;
  createUserFullname: string;
  statusNameEn: string;
  statusName: string;
  typeId: string;
  id: string;
  ticketCode: string;
  statusNameRu: string;
  createDate: string;
}